python/src/auth/requirements.txt
astroid==3.0.2
blinker==1.7.0
click==8.1.7
dill==0.3.8
Flask==3.0.2
Flask-MySQLdb==2.0.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
mysqlclient==2.2.2
parso==0.8.3
platformdirs==4.2.0
PyJWT==2.8.0
pylint==3.0.3
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
Werkzeug==3.0.1
